+++ R/run_query.R
@@ -15,17 +15,23 @@ run_query <- function(query, conn = last_connection()) {
   if (missing(query)) return(last_connection())
   stopifnot(is(conn, 'JDBCConnection')) 
   stopifnot(is.character(query))
-  # Attempts to connect with the database.  If it fails, it will reload the connection once and try again.
-  tryCatch({res <- RJDBC::dbSendQuery(conn, query)
-            RJDBC::fetch(res, n = -1)},
-           error = function(c) {
-             message("* Error Occurred, Reloading Reload Connection...")
-             tryCatch( {
-               suppressMessages(conn <- new_connection());
-               res <- RJDBC::dbSendQuery(conn, query);}
-               ,error = function(c) {
-                 message('* Reloading the Connection did not fix the Error.'); 
-                 message(c) })
-           }
+  # Attempts to connect with the database.  If it fails, it will reload the connection and try again.
+  tryCatch({
+      res <- RJDBC::dbSendQuery(conn, query)
+      RJDBC::fetch(res, n = -1)
+    },
+    error = function(c) {
+      message("* Error Occurred, Reloading Connection...")
+      tryCatch({
+          suppressMessages(conn <- postgresql_connection(getOption('database.yml'), strict = FALSE));
+          res <- RJDBC::dbSendQuery(conn, query);
+          RJDBC::fetch(res, n = -1);
+        },
+        error = function(c) {
+          message('* Reloading the Connection did not fix the Error.'); 
+          message(c) 
+        }
+      )
+    }
   )
 }